pyproject.toml
[project]
name = "dailyknowledge"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "google-generativeai>=0.8.0",
    "markdown>=3.5.0",
    "openai>=1.0.0",
    "premailer>=3.10.0",
    "python-dotenv>=1.0.0",
]
